Move seeding to it's own file, sync before refresh
Seeding was originally handled as part of the handler, as this functionality
may want to be used other ways in the future, it's been broken out to allow for
that. This also moves the sync before the refresh background jobs are
instantiated to ensure that there are keys already in the repo prior to
syncing. Print statements in key_refresh.py have been replaced with the logger
for consistency.