Make seeding on init optional but default to true
Give system the ability to not seed using the built-in python method. This is a
good option for when you GPG keys are in binary format as the python-gnupg
libary does not support loading binary keys. To align with the current seeding
configuration it is set to enabled by deault and can be disabled with the
environment variable "SEED_ON_INIT=0".