Fix issue with CSR subject validation being too lenient
CSR subject validation is too lenient allowing for unverified data to exist in
the final Certificate. To prevent that only check and allow the first Email
inside a subject as well as limit the types of OIDs that are allowed in the
Subject. Support for using RFC822 in the SAN is now also supported as
Email is now deprecated in Certificates. Also removed DNS names from the final
certificate as they were not needed ever.

Closes #42
Closes #41