Configurable timeout for PGP signed messages.
Allow an operator to set the timeout on PGP Signed Messages.

This defaults to 5 seconds, which is generally more than enough time for a
request to be signed and then sent to the server to begin processing. Allowing
this to be configurable will allow operators to teak the timeout based on
real-world use within their infrastructure.